SDK-1030: Move Anchor tests from ActivityTests to AnchorTests; extract common method to TestTools
